backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
httpx==0.28.1
beautifulsoup4==4.13.3
lxml==5.3.1
pydantic==2.11.3
pydantic-settings==2.8.1
PyJWT==2.10.1
slowapi==0.1.9
google-genai==1.10.0
python-multipart==0.0.20
structlog==25.1.0
python-dotenv==1.1.0
tenacity==9.1.2
